completely refactored browse page
SearchBar, BrowseMovieList, and MovieCard are now all siblings in BrowsePage, instead of all being in a MovieList monolith.
