+++ src/components/index.js
@@ -1,4 +1,5 @@
 export { default as AccountNav } from './AccountNav';
+export { default as BrowseMovieList } from './BrowseMovieList';
 export { default as ChatWindow } from './ChatWindow';
 export { default as FeedbackToast } from './FeedbackToast';
 export { default as GenreList } from './GenreList';
@@ -7,10 +8,10 @@ export { default as Login } from './Login';
 export { default as MovieCard } from './MovieCard';
 export { default as MovieFeed } from './MovieFeed';
 export { default as MovieFeedItem } from './MovieFeedItem';
-export { default as MovieList } from './MovieList';
 export { default as Navbar } from './Navbar';
 export { default as OAuth } from './OAuth';
 export { default as SearchBar } from './SearchBar';
 export { default as Signup } from './Signup';
 export { default as TrailerModal } from './TrailerModal';
+export { default as UserCheckboxList } from './UserCheckboxList';
 export { default as Watchlist } from './Watchlist';
