+++ src/css/MovieCard.css
@@ -1,16 +1,15 @@
 @import './style-variables.css';
 
-:root {
-  --modalheight: 82vh;
-  --modalwidth: 85vw;
-}
-
 .movie-card {
+  --modalheight: 82vh;
+  --modalwidth: 88vw;
+
   display: flex;
   justify-content: center;
-  padding: 3vh 0px;
+  padding: 2vh 0px;
   z-index: 10;
   position: fixed;
+  top: var(--nav-height);
   width: 100%;
   height: 100%;
   font-family: var(--fontfamily);
@@ -229,22 +228,21 @@
   border-color: var(--red);
 }
 
-
 /* ======================================================= */
 /* Watchlist Button*/
 .watchlist {
-    position: absolute;
-    right: 10px;
-    bottom: 10px;
-    border: none;
-    margin: 0;
-    text-decoration: none;
-    background-color: none;
-    color: var(--purple)
+  position: absolute;
+  right: 10px;
+  bottom: 10px;
+  border: none;
+  margin: 0;
+  text-decoration: none;
+  background-color: none;
+  color: var(--purple);
 }
 
 .watchlist:hover {
-  color: var(--red)
+  color: var(--red);
 }
 
 /* ======================================================= */
