+++ src/css/main.css
@@ -12,7 +12,7 @@
 /* =========================== MAIN =========================== */
 
 body {
-  margin-top: 55px;
+  margin-top: var(--nav-height);
   background-color: var(--darkgrey);
   font-family: var(--fontfamily);
   font-size: 20px;
