+++ src/css/style-variables.css
@@ -7,6 +7,8 @@
   --darkgrey: #23262e;
   --offwhite: #f5f8e8;
   --popcornyellow: #ffe695;
+
+  --nav-height: 65px;
   --fontfamily: Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue',
     sans-serif;
 }
